machine/rp2: merge common usb code
